+++ packages/cli/src/config/config.ts
@@ -572,7 +572,7 @@ export async function loadCliConfig(
           'https://dashscope.aliyuncs.com/compatible-mode/v1/',
           'https://dashscope-intl.aliyuncs.com/compatible-mode/v1/',
         ],
-        modelNames: ['qwen3-coder-plus'],
+        modelNames: ['Llama-4-Maverick-17B-128E-Instruct-FP8'],
         template:
           'SYSTEM_TEMPLATE:{"name":"qwen3_coder","params":{"is_git_repository":{RUNTIME_VARS_IS_GIT_REPO},"sandbox":"{RUNTIME_VARS_SANDBOX}"}}',
       },