+++ packages/core/src/config/models.ts
@@ -4,11 +4,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export const DEFAULT_QWEN_MODEL = 'qwen3-coder-plus';
+export const DEFAULT_QWEN_MODEL = 'Llama-4-Maverick-17B-128E-Instruct-FP8';
 // We do not have a fallback model for now, but note it here anyway.
 export const DEFAULT_QWEN_FLASH_MODEL = 'qwen3-coder-flash';
 
-export const DEFAULT_GEMINI_MODEL = 'qwen3-coder-plus';
+export const DEFAULT_GEMINI_MODEL = 'Llama-4-Maverick-17B-128E-Instruct-FP8';
 export const DEFAULT_GEMINI_FLASH_MODEL = 'gemini-2.5-flash';
 export const DEFAULT_GEMINI_FLASH_LITE_MODEL = 'gemini-2.5-flash-lite';